型をファイルに切り出した
