+++ src/features/stages/components/Instruction.tsx
@@ -1,11 +1,6 @@
 import React, { useCallback } from 'react';
 import Image from 'next/image';
-import {
-  CONDITION,
-  Condition,
-  MODE,
-  Mode,
-} from '../../../pages/stages/[stage]';
+import { CONDITION, Condition, MODE, Mode } from '../types';
 
 type Props = {
   condition: Condition;
